requirements.txt
click==6.6
Flask==0.11.1
Flask-Login==0.3.2
flask-peewee==0.6.7
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.1
pandas==0.18.1
patsy==0.4.1
peewee==2.8.1
PyMySQL==0.7.4
python-dateutil==2.5.3
pytz==2016.4
scikit-learn==0.17.1
scipy==0.18.0
six==1.10.0
sklearn==0.0
statsmodels==0.8.0rc1
Werkzeug==0.11.10
wtf-peewee==0.2.6
WTForms==2.1